Thanks to @hkopp for the help!
Moved all of the project entries from the current website over to the new one.
There are a lot of incorrectly set dates (which need manual fixing) and some missing information/outdated things.
We also need to move a bunch of stuff to the archives.

This commit also changes the fact that 'project' is now a category instead of the 'isproject' variable, which was kinda poorly designed.